Moved cli helpers into cli
